Refactor /mappings/common into /common
