+++ internal/common/map.go
