+++ internal/common/map_test.go
