--- internal/mappings/common/errors.go
@@ -1,9 +1,0 @@
-package common
-
-const (
-	// IncorrectFormat is the error text for an incorrect format in a mapping.
-	IncorrectFormat = "Incorrect format (in '%s')"
-
-	// MissingValue is the error text for a missing value in a mapping.
-	MissingValue = "Missing value (in '%s')"
-)
